refactor: chunkArray를 userService에서 공용 유틸리티로 추출
